requirements.txt
transformers==4.40.0
torch
Pillow
gradio
matplotlib
numpy